% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_logLik.R
\name{calculate_logLik}
\alias{calculate_logLik}
\title{Calculate log-likelihood for acceptance ratio in MCMC scheme.}
\usage{
calculate_logLik(tposInd, obsInd, alphaLeft, alphaRight, etaPart, betaPart)
}
\arguments{
\item{tposInd}{A n x 1 vector indicating whether the event was observed before last follow-up.}

\item{obsInd}{A n x 1 vector indicating whether the event was observed after follow-up started.}

\item{alphaLeft}{A n x p matrix of the left time point fixed coefficients times the covariates.}

\item{alphaRight}{A n x p matrix of the right time point fixed coefficients times the covariates.}

\item{etaPart}{A n x (nKnots + 2) matrix of the spline coefficients times cubic spline values.}

\item{betaPart}{A n x q matrix of the estimated genetic coefficients times the SNP data.}
}
\value{
A scalar value summing the log likelihood of the left and right time points.
}
\description{
Calculate log-likelihood for acceptance ratio in MCMC scheme.
}
