% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThresholdBeta.R
\name{getThresholdBeta}
\alias{getThresholdBeta}
\title{Type I error threshold for estimated genetic effects.}
\usage{
getThresholdBeta(q, estBeta, causal, T1E)
}
\arguments{
\item{q}{Total number of SNPs being fine-mapped.}

\item{estBeta}{Estimated effect sizes.}

\item{causal}{True causal variants.}

\item{T1E}{Desired Type 1 Error rate.}
}
\value{
A single value indicating the threshold for beta estimate.
}
\description{
Type I error threshold for estimated genetic effects.
}
