% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThresholdPIP.R
\name{getThresholdPIP}
\alias{getThresholdPIP}
\title{Type I error threshold for estimated posterior inclusion probabilities.}
\usage{
getThresholdPIP(p, pip, causal, T1E)
}
\arguments{
\item{pip}{A 1 x q vector of the estimated posterior inclusion probabilities.}

\item{causal}{True causal variants.}

\item{T1E}{Desired Type 1 Error rate.}

\item{q}{Total number of SNPs being fine-mapped.}
}
\value{
A single value indicating the threshold for beta estimate.
}
\description{
Type I error threshold for estimated posterior inclusion probabilities.
}
