% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThresholdPval.R
\name{getThresholdPval}
\alias{getThresholdPval}
\title{Type I error threshold for frequentist p-value.}
\usage{
getThresholdPval(p, pvals, causal, T1E)
}
\arguments{
\item{pvals}{A 1 x q vector of the calculated p-values.}

\item{causal}{True causal variants.}

\item{T1E}{Desired Type 1 Error rate.}

\item{q}{Total number of SNPs being fine-mapped.}
}
\value{
A single value indicating the threshold for beta estimate.
}
\description{
Type I error threshold for frequentist p-value.
}
