% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icbvs_gen_dat.R
\name{icbvs_gen_dat}
\alias{icbvs_gen_dat}
\title{Generating interval censored data.}
\usage{
icbvs_gen_dat(
  n,
  gMat,
  noCausal,
  rho,
  ES,
  bhFunInv = function(x) {
     x
 },
  obsTimes = 1:5,
  windowHalf = 0.1,
  trueEta = c(-1, 0)
)
}
\arguments{
\item{n}{Total number of observations.}

\item{gMat}{A n x q matrix of the genetic data.}

\item{noCausal}{Number of causal genetic variants.}

\item{rho}{Common pairwise correlation parameter.}

\item{ES}{Effect size of a causal genetic variant.}

\item{bhFunInv}{The inverse of the baseline hazard function.}

\item{obsTimes}{Vector of the intended observation times.}

\item{windowHalf}{The amount of time before or after the intended obsTimes that a visit might take place.}

\item{trueEta}{A vector of true effect sizes for the fixed effects.}
}
\value{
A list with the elements:
\item{outcomeDat}{List containing the simulated left and right times.}
\item{trueBeta}{A 1 x q vector of the true genetic effects used to simulate the data.}
\item{xMat}{A n x p matrix of the simulated covariate data.}
}
\description{
Generating interval censored data.
}
