% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_step_hs.R
\name{mcmc_step_hs}
\alias{mcmc_step_hs}
\title{One step of the MCMC scheme using the horseshoe prior specification.}
\usage{
mcmc_step_hs(
  leftDmat,
  rightDmat,
  tposInd,
  obsInd,
  zetaCurr,
  alphaCurr,
  etaCurr,
  betaCurr,
  sigmaAprop,
  sigmaEprop,
  sigmaBprop,
  sigmaAprior,
  sigmaEprior,
  sigmaBprior,
  gammaCurr,
  removeProb,
  addProb,
  alphaPriorMean,
  etaPriorMean,
  A = 1,
  subPct = 0.5
)
}
\arguments{
\item{leftDmat}{Left design matrix}

\item{rightDmat}{Right design matrix}

\item{tposInd}{A n x 1 vector indicating whether the event was observed before last follow-up.}

\item{obsInd}{A n x 1 vector indicating whether the event was observed after follow-up started.}

\item{zetaCurr}{Current zeta values.}

\item{alphaCurr}{Current alpha values.}

\item{etaCurr}{Current eta values.}

\item{betaCurr}{Current beta values.}

\item{sigmaAprop}{Hyperparameter for the standard deviation of the proposal distribution for the spline coefficients.}

\item{sigmaEprop}{Hyperparameter for the standard deviation of the proposal distribution for the fixed effect coefficients.}

\item{sigmaBprop}{Hyperparameter for the standard deviation of the proposal distribution for the genetic effect coefficients.}

\item{sigmaAprior}{Hyperparameter for the standard deviation of the prior distribution for the spline coefficients.}

\item{sigmaEprior}{Hyperparameter for the standard deviation of the prior distribution for the fixed effect coefficients.}

\item{sigmaBprior}{Hyperparameter for the standard deviation of the prior distribution for the genetic effect coefficients.}

\item{gammaCurr}{Current gamma values.}

\item{removeProb}{Probability of removing another causal variant in selection scheme.}

\item{addProb}{Probability of adding another causal variant in selection scheme.}

\item{alphaPriorMean}{A 1 x (nKnot + 2) vector of the spline coefficients.}

\item{etaPriorMean}{A 1 x p vector of the fixed effects.}

\item{A}{Hyperparameter in rate parameter in gamma prior. Default set to 1.}

\item{subPct}{A single value from 0 to 1 indicating the percentage of genetic variants to update each iteration of the MCMC. Default set to 0.5.}
}
\value{
A n x q matrix of genetic data.

A list with the elements:
\item{betaCurr}{Matrix of current beta values.}
\item{alphaCurr}{Matrix of current alpha values.}
\item{etaCurr}{Matrix of current eta values.}
\item{gammaCurr}{Matrix of current gamma value}
\item{accMat}{A (B\emph{burnIn) x (q}2 + (nKnot + 2) + p + 1) matrix containing indicators for whether the proposal was accepted (= 1) or not (=0).}
}
\description{
One step of the MCMC scheme using the horseshoe prior specification.
}
