% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_step_sparse.R
\name{mcmc_step_sparse}
\alias{mcmc_step_sparse}
\title{One step of the MCMC scheme using SS-Sparse prior specification.}
\usage{
mcmc_step_sparse(
  leftDmat,
  rightDmat,
  tposInd,
  obsInd,
  piCurr,
  alphaCurr,
  etaCurr,
  betaCurr,
  sigmaAprop,
  sigmaEprop,
  sigmaBprop,
  sigmaAprior,
  sigmaEprior,
  sigmaBprior,
  gammaCurr,
  removeProb,
  addProb,
  alphaPriorMean,
  etaPriorMean
)
}
\arguments{
\item{leftDmat}{The design matrix for the left times.}

\item{rightDmat}{The design matrix for the right times.}

\item{tposInd}{A n x 1 vector indicating whether the event was observed before last follow-up.}

\item{obsInd}{A n x 1 vector indicating whether the event was observed after follow-up started.}

\item{piCurr}{A vector containing of the current pi values.}

\item{alphaCurr}{A matrix of current alpha values.}

\item{etaCurr}{A matrix of the current eta values.}

\item{betaCurr}{A matrix of the current beta values.}

\item{sigmaAprop}{Hyperparameter for the standard deviation of the proposal distribution for the spline coefficients.}

\item{sigmaEprop}{Hyperparameter for the standard deviation of the proposal distribution for the fixed effect coefficients.}

\item{sigmaBprop}{Hyperparameter for the standard deviation of the proposal distribution for the genetic effect coefficients.}

\item{sigmaAprior}{Hyperparameter for the standard deviation of the prior distribution for the spline coefficients.}

\item{sigmaEprior}{Hyperparameter for the standard deviation of the prior distribution for the fixed effect coefficients.}

\item{sigmaBprior}{Hyperparameter for the standard deviation of the prior distribution for the genetic effect coefficients.}

\item{gammaCurr}{A matrix of the current gamma values.}

\item{removeProb}{Probability of removing another causal variant in selection scheme.}

\item{addProb}{Probability of adding another causal variant in selection scheme.}

\item{alphaPriorMean}{A 1 x (nKnot + 2) vector of the spline coefficients.}

\item{etaPriorMean}{A 1 x p vector of the fixed effects.}
}
\value{
A n x q matrix of genetic data.

A list with the elements:
\item{betaCurr}{Matrix of current beta values.}
\item{alphaCurr}{Matrix of current alpha values.}
\item{etaCurr}{Matrix of current eta values.}
\item{gammaCurr}{Matrix of current gamma value}
\item{accMat}{A (B\emph{burnIn) x (q}2 + (nKnot + 2) + p + 1) matrix containing indicators for whether the proposal was accepted (= 1) or not (=0).}
}
\description{
One step of the MCMC scheme using SS-Sparse prior specification.
}
