% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate sample interval-censored data.}
\usage{
sim_data(
  seed = 0,
  n,
  q,
  bhFunInv,
  obsTimes,
  windowHalf,
  trueEta = NULL,
  trueBeta,
  gMat
)
}
\arguments{
\item{seed}{Random seed.}

\item{n}{Total number of observations.}

\item{q}{Total number of SNPs.}

\item{obsTimes}{Vector of the intended observation times.}

\item{windowHalf}{The amount of time before or after the intended obsTimes that a visit might take place.}

\item{trueEta}{A vector of true effect sizes for the fixed effects.}

\item{trueBeta}{A vector of true effect sizes for the random effects.}

\item{gMat}{A n x q matrix of the genetic data.}

\item{bnFunInv}{The inverse of the baseline hazard function.}
}
\value{
A list containing the outcome times, covariate data, and genetic data.
}
\description{
Simulate sample interval-censored data.
}
