% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gMat.R
\name{sim_gMat}
\alias{sim_gMat}
\title{Generate simulated genetic data set}
\usage{
sim_gMat(n, q, rho, mafVec)
}
\arguments{
\item{n}{Total number of observations.}

\item{q}{Total number of SNPs.}

\item{rho}{Common pairwise correlation parameter.}

\item{maxMAF}{Maximum minor allele frequency for mean parameter vector.}
}
\value{
A n x q matrix of genetic data.
}
\description{
Generate simulated genetic data set
}
\examples{
sim_gMat(1000, 100, 0.1, .05)
}
